@@ -10,16 +10,18 @@ dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.6
-  badges: 2.0.2
-  flutter_svg: 2.0.9
-  font_awesome_flutter: 10.6.0
-  google_fonts: 6.1.0
+  vibration: 1.9.0
+  badges: 3.1.2
+  flutter_svg: 2.0.10+1
+  font_awesome_flutter: 10.7.0
+  google_fonts: 6.2.1
+  device_info_plus: ^10.1.0
   awesome_ripple_animation: 2.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^3.0.0
+  flutter_lints: 4.0.0
 
 flutter:
   uses-material-design: true
